Retain an AnimationClock per-document instead of per-page
Since each document coarsens timestamps differently, having a clock
for each makes more sense than having one per the all page and then
coarsening when retrieving the time.

Coarsening on the getter was causing a performance regression, so
coarsening at the start of the frame when setting the clock fixes it,
but requires a separate clock per document.

We still keep the page clock for things like SVG images and intersection
stability, though that can be removed in the future.

This also matches the spec more closely, where time is coarsened
when updating the rendering rather than in DocumentTimeline.currentTime.

Bug: 1496136
Change-Id: Ifbd4ca03471232598585db9ee38208c00bcead99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4981913
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1217900}
NOKEYCHECK=True
GitOrigin-RevId: f96393629da7d901caaec710af6710c0a573e47c