Update no hover during scroll tests
Speculative deflake of no-hover-during-js-scroll test.  This test
periodically flakes on Windows when run in threaded virtual suite
with percent based scrolling enabled. When it flakes, the result
is a timeout. Switching to a mouse click instead of a mouse move and
waiting on the pointer up ensures that the window has input focus and
will process hover state changes. Switches from the gpubenchmarking
API to test-driver.

Bug: 1481219
Change-Id: I4f0e57d54b967984b51b6eec406e814a242612d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4985262
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218195}
NOKEYCHECK=True
GitOrigin-RevId: d3bf7ee76bf8f9020a38004b46f9258a2079602e