[SAA] Inherit the context's storage-access bit for WebSockets
This plumbs the has_storage_access bool from the WebSocket's
ExecutionContext to the network service machinery, so that a
Document's or Worker's Storage Access API opt-in is respected for
WebSocket connections. (Note that this bit is untrusted; a compromised
renderer cannot gain unauthorized access to unpartitioned cookies by
lying with this bit.)

This behavior is currently unspecified by the Storage Access API spec,
but we will engage there to specify it. In the meantime, web
developers expect the API to work this way (see
https://crbug.com/947413#c20).

I plan to add WPTs for this in a followup, but I'd like to land this
before branch cut if possible.

Bug: 1496832
Change-Id: I29e0e4b3a2fc6fb3faed8c80954aaee14499e449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4982327
Reviewed-by: Johann Hofmann <johannhof@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218202}
NOKEYCHECK=True
GitOrigin-RevId: 2145062963b5882d12e999b9df114732affb5304