Workaround for 3PCB when rendering sandboxed media
When 3PCB is enabled, access-controlled media URLs that
are served with "Content-Security-Policy: sandbox" will be
broken when it is loaded as the top-level URL.

This includes static media types where there is little privacy
or security risk to allowing cookies to be send for the data
request when the media is being loaded as the top-level document.

The CL patches this by overriding Document::SiteForCookies in these
specific cases. To be cautious, this is only done in the
MediaDocumentParser.

Bug: 1475014
Change-Id: I5b14785394ce6c230ab2a970097582b2b07c5c1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4980051
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218420}
NOKEYCHECK=True
GitOrigin-RevId: fcbae4f06caaf56d0cea2e42d78ca651277b8342