[sticky] Don't invalidate sticky constraints that don't exist yet.
InvalidatePaintForStickyDescendants may be call in the middle of
layout when we delay scroll offset clamping (scroll offsets will get
set in the middle of layout).

The layout-box will be layout-clean - however the sticky descdendants
might not have any sticky-pos constraints yet.

Check if we have been enqueued for a sticky update, and skip the
paint invalidation step (it'll happen later when we regenerate the
sticky constraints).

Bug: 1498297, 1498401
Change-Id: I7f0f4cd86203b845cd37a015cb7bd485773494f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4997393
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218508}
NOKEYCHECK=True
GitOrigin-RevId: e224cc7d98f4fcdddd96ca762d66b5de667dde6a