[AssistData] Add histograms to track when max timeout or nodes is exceeded
This CL continues the work to add more tracking for accessibility
snapshots on Android, and their failure rates / failure reasons.

With this CL we add three more types of failures to the enum of possible
AXTree snapshot failure reasons. We consider the case where the max
number of nodes is met, the timeout is reached, or both. We create a new
ax_error_types.h file with the possible types of serialization errors.
We add a new optional parameter to the SerializeChanges method of the
ax_tree_serializer, which when populated will return the error type of
the given serialization call (if one occurs). The possible error types
are max nodes reached, timeout reached, or both. We treat the enum as
a bitmask for convenience.

We include a unit test in the Java-side code that makes a tree much
larger than the max node limit and verifies that the snapshot only
includes a partial tree.

Bug: b/303655535
Change-Id: Ie4fc5e24e328a93f284dc469d2acf9cbcd59168d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4974049
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1219060}
NOKEYCHECK=True
GitOrigin-RevId: f5a7cd015b9cac36f642a621605baa16cef1e761