Add explicitness booleans for PixelsAndPercent fields
Per css-values-4 spec we should preserve the 0 unit after
simplifications, but we should not preserve it if it isn't in the
expression (e.g., min(10%, 10rem) should not be min(10% + 0px, 0% + 160px).

Currently, there is no way to indicate that in
PixelsAndPercent, so, two new bools were added in order to indicate
if the pixels and percent are explicitly used in expression.

The optionals are not used in order not to change the api of the struct, as it's read in many places, but set in few.
So, not to change the read api, two booleans have been preferred over the optionals.

Bug: 1050968
Change-Id: If5cde5126eb9c95e395045e1b6f496820f8a80e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5000667
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1219516}
NOKEYCHECK=True
GitOrigin-RevId: 86238c5ce1351cf2f49a2c68dbe3ab984728a2f7