content: Skip deferring main frame updates for RenderDocument
With RenderDocument + compositor reuse, we swap in a new frame which
reuses the previous frame's compositor. The compositor needs to start
with BeginMainFrames paused to avoid flashing a blank frame. This is
done in WebViewImpl::DidAttachLocalMainFrame().

The same functionality also exists in DidChangeRootLayer() which is
dispatched when the Document is being detached. This is incorrect if
the compositor has already been passed to the new widget with
RenderDocument + compositor reuse. It is also unnecessary because it
will be done in DidAttachLocalMainFrame().

Bug: 1464791
Change-Id: If2cac3f5585357a01df45bce9318a382346440bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994379
Auto-Submit: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1219560}
NOKEYCHECK=True
GitOrigin-RevId: c71681e3aa47bf6a59223079af871317c739f198