[wpt-importer] Only fill in results for incomplete builds
The importer's expectations updater inherits results for "missing
builds" (currently defined as zero unexpected failures). This usually
works in practice because version-specific failures are very rare, but
it would be better to not conflate incomplete with successful builds
(the latter should not be filled). Use the `BuildResolver(...)`
definition of an incomplete build when deciding whether to fill in
expectations.

Bug: 1481304
Change-Id: Id0de795520db63a76f19e119529be96ddcc30a41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5005468
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1219756}
NOKEYCHECK=True
GitOrigin-RevId: 65e5d48d4ba4c310275c2f74f90142f9d1458322