Repurpose and redo the AdClickNavigation use counter
The existing AdClickNavigation UseCounter is recorded in the renderer
at navigation initiating time. There's a flaw in the UseCounter
recording architecture, such that UseCounters recorded from
cross-origin subframes could get lost if the page gets navigated
away soon, because the UseCounters are queued in the renderer and
flushed periodically, and metrics from subframes won't be flushed
when the top-level page navigates away.

This CL switches to log it in the browser side at the navigation IPC
boundaries. Besides, we figured it's more desirable to record for
main-frame navigation only. Thus, AdClickNavigation is repurposed
to AdClickMainFrameNavigation (deprecate the old one, and introduce
the new UseCounter).

Bug: 1496605
Change-Id: I0b7f858d7319dbbcf65ddc2e6c5cc8c5bfcf87fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4983085
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1219772}
NOKEYCHECK=True
GitOrigin-RevId: 65002fb897b307077d8057a0daf15cb647a8c6c5