[Files]: Adding cutoff days as a parameter to getRecentFiles.
The recent model already had the ability to use cutoff time, but only
in tests. We are making it available explicitly to allow recent view
show results for other than the default, 30 days, modified time. In
JavaScript this is added as a cutoffDays parameter that is converted
to TimeDelta in Days. In the C++ any TimeDelta, be it days or seconds,
is considered valid.

The non-default modified time is only used when searching, not in the
default Recents view.

Bug: b:307455066
Change-Id: Ia3b58bc24995d7bbef3b161420bff57f3cc3af98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4988981
Commit-Queue: Bo Majewski <majewski@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220093}
NOKEYCHECK=True
GitOrigin-RevId: e5a6dea9e5b52530d89964fb21e87eee49ceb8ff