[LCPP] preloading LCP candidate images with loading=lazy
This CL implements the preloading behavior to LCP candidate images,
which have loading="lazy" attributes behind the flag.

In theory loading="lazy" is added to the image located below the fold
for the loading start timing optimization. If the LCP image has the
attribute, that would negatively affect the metric.

This CL updates the preload scanner to create a preload request
regardless of whether the loading="lazy" attribute exists or not when
the image element is in the LCPP element locator.

In the future, we have an idea that extending this behavior to the JS
based custom lazy loading, but this CL starts from the native lazy
loading.

Change-Id: If4bc9c7117e08cda31f1e5ff01f44ac038448ba6
Bug: 1498777
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4996930
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220145}
NOKEYCHECK=True
GitOrigin-RevId: a8eaaeee0846329ae0d8e699c7a306225939d2b1