Keep track of sourcemaps/source urls for constructed stylesheets
When reporting stylesheets to the inspector, the sourceMappingURL or
sourceURL is parsed from the stylesheet text. For constructed
stylesheets, this text isn't available, and the inspector rebuilds it
from the rules, loosing the sourceMappingURL comment. To avoid this, we
preserve the text.

Bug: 1313674
Change-Id: I43f6c5f80acb296568e4302d94d3a19e812cc304
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4955818
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220158}
NOKEYCHECK=True
GitOrigin-RevId: b2d809e15aae45762e7c66232112a6b202a795b1