[Async Clipboard API] Remove unsanitized write permissions.
The definition of sanitization for HTML format write was changed
to write well-formed HTML document instead of a fragment when
`kClipboardWellFormedHtmlSanitizationWrite` flag is enabled.
This was done to address the information loss with the sanitized
fragment.

- The behavior with the flag changed from passing through unaltered
HTML i.e. unsanitized, to passing through well-formed HTML document
that has been processed by the DOM parser.
- The strict `CLIPBOARD_READ_WRITE` permission check was necessary
for writing unsanitized HTML, but is unnecessary for writing
lightly-sanitized HTML.

The `PermissionType` that gates the action of writing HTML format to
the clipboard is being changed from `CLIPBOARD_READ_WRITE`
to `CLIPBOARD_SANITIZED_WRITE`. This has the notable effect of
avoiding a permission prompt in some situations.

The behavior for web custom formats write permission remains
unchanged.

Bug: 1268679

Change-Id: I992fb6b17f2f32bb183c88d9dd94e8bce2390a35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4988632
Commit-Queue: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Sanket Joshi <sajos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1220412}
NOKEYCHECK=True
GitOrigin-RevId: f4f1d6a3da26b83b8c4d9fd2fc64d84b56516221