IndexedDB: simplify by removing FakeLevelDBFactory
The LevelDBFactory interface existed to facilitate testing and had
two implementations: DefaultLevelDBFactory and FakeLevelDBFactory.
FakeLevelDBFactory is not used much, and has almost no actual
logic, so it's easy to replicate its functionality in the couple of
places where it was being used.

Thus it's possible to get rid of the LevelDBFactory interface since
we don't need multiple implementations. DefaultLevelDBFactory is
kept around for now, but is renamed to LevelDBFactory. At this point
its purpose is to hold onto IDB-specific LevelDB settings, but it
is also subject to further refinement in the near future.

Bug: 1474996
Change-Id: I41a1a575508fd553d347e61d838b6fc2acb822ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5003219
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220422}
NOKEYCHECK=True
GitOrigin-RevId: cbc7df2cbea2125c6dc4ffeac758dac8a6ea7a61