[RemoveLegacy] (Re)-Allow anonymous blocks to be containing blocks.
We did this as part of LayoutNG originally to correctly position OOFs
within split inlines.
As part of the "block-in-inline" project we removed inline-splitting!

As such we can re-allow anonymous blocks to be containing blocks. This
happens when an inline (<span>) is a container for an OOF, but we
position it relative to an anonymous wrapping block (created for
block-in-inline purposes for example).

Behind LayoutNewContainingBlock flag.

Bug: 1229581
Change-Id: I4e3e38ae490f082f1204505aeeb2e973f0fa6c33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5005819
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220512}
NOKEYCHECK=True
GitOrigin-RevId: ebca14d429886b5775ad959067b8f97f2f1413e5