Fix EventSender tests that will break under isolated sandboxed iframes.
This CL updates a number of WPT tests that rely on EventSender, and that
will break when enable-isolated-sandboxed-iframes is enabled by default.
EventSender only works within a single renderer process, but the
affected tests will create situations where a parent cross-process to
a child will try and send and event using EventSender, and the event
will fail to reach the child.

The approach taken in this CL is to have the parent notify the child
via postMessage when an event is required, and the child will invoke the
event locally with EventSender. The revised tests will work correctly
regardless of whether enable-isolated-sandboxed-iframes is enabled or
not.

Bug: 550122
Change-Id: I7a86889ae3effa49a55fac149afda49c2f13ee5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5007016
Commit-Queue: W. James Maclean <wjmaclean@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220518}
NOKEYCHECK=True
GitOrigin-RevId: d35b603ae7e11858e9fa13acdae8ebbdc6e48ed4