Replace event report windows field with trigger specs in DevTools protocol
Using the default trigger specs for the corresponding source type until
the SourceRegistration structure itself contains a field for the specs.

The DevTools protocol structure mirrors that of the JSON registration:
a list of specs, each of which is identified by a list of trigger data
values and described by an event report windows structure.

Bug: 1499891
Change-Id: I412338e2ec950c8aae272561295c936b97d6f637
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5006002
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220627}
NOKEYCHECK=True
GitOrigin-RevId: 93ea5faaffebcf05268cae9c8a8afcd68af5492b