[view-transition] Reland the VT/BFCache fix
b/302490197 was introduced by:
- https://crrev.com/c/4702023: calling MoveToFront on the old
  ViewAndroid. Supposed to be called on the new ViewAndroid.
- https://crbug.com/1492600: BFCache accidentally stores a
  non-history-traversable page, when the navigation is made from an
  OOPIF via window.top.location.replace, with coop=same-site response
  header.

This CL corrects the behavior of https://crrev.com/c/4702023. It adds
the test to make sure that the main frame's ViewAndroid is always at
top-most. It also adds a test to reproduce the behavior in
https://crbug.com/1492600.

The feature flag kInvalidateLocalSurfaceIdPreCommit is re-enabled by
this CL. The changes to re-enable the flag are:
- components/viz/common/features.cc flag re-enabled.
- render_frame_host_impl_browsertest.cc `DISABLED_` macro.
- testing/variations/fieldtrial_testing_config.json FT param removed.

PTAL at the original CL https://crrev.com/c/4702023, especially its CL
description, which captures the motivation of the original change.

The feature flag will be enabled in a follow-up CL.

Bug: 1492600, 1453638, 1415340, 1431273, 1361305, 1495157
Low-Coverage-Reason: OTHER RWHVA's existing behaviour is not tested.
Change-Id: I006a6d6655855bccbbdfe36bb130fe7a529d3de6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4979712
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: William Liu <liuwilliam@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220628}
NOKEYCHECK=True
GitOrigin-RevId: b4cb61e9cbd561b171f6c8274672cbcb18e5718d