[A11y] Keep all slots and their children in a11y tree
This restores the logic for slots inside of
ComputeAccessiblityIsIgnoredButIncludedInTree() to what it was previously. It now again that all slots and their accessible children
are included in the tree, regardless of whether they support assignment.
This makes it possible to properly clean things up even after an
assignment changes, or is gained or lost.

This change was pulled out of CL:4873421, and without it, a number
of tests hit checks in that CL.

Bug: none
Change-Id: I5f8a37bf92b1976b1daa5f132241ae0da4d04deb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5006618
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220672}
NOKEYCHECK=True
GitOrigin-RevId: c571017136cfabe26b25b621a0c750e0e8233737