Sort all the document markers
To improve the efficiency of marker painting, it is desirable to
keep all markers sorted, including overlapping markers. This allows
us to avoid sorting markers at paint time, over and over again.

The existing UnsortedDocumentMarkerListEditor code is updated to
keep the markers sorted on StartOffset().

Additional testing for overlapping markers, particularly nested
markers, is added. Duplicate methods between the sorted and unsorted
marker editors are rationalized. Calling code is updated.

This CL does not change the code in DocumentMarkerController to
take advantage of the sort. That will be added in a follow on CL.

Bug: 1442067
Change-Id: I6c659d703c76ab84fdfe72c2ac3f45fe510b6298
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5004942
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220683}
NOKEYCHECK=True
GitOrigin-RevId: 91466bf9233325f129bfff8c14d32f418ee0e8c7