Add [Default] to some Extensible mojo enums
Mojo Extensible enums should have a [Default] entry to use when
an out-of-range value is supplied. This CL resolves a couple of
cases that have been allowlisted.

1. chrome_cleaner is gone it can be deleted.

2. components/digital_goods/mojom/digital_goods.mojom

All enums have a reasonable default (kError, kUnknown) for
out-of-range values.

3. third_party/blink/public/mojom/page_state/page_state.mojom

Page state is persisted - the default is Auto so this is a
reasonable default. It also is unlikely that newer values will be
exposed to older interfaces if any more fields are added.

Bug: 1316161
Change-Id: I3a43a08a17d8feeeefe1041af6823a36e8125519
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4984487
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: W. James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220688}
NOKEYCHECK=True
GitOrigin-RevId: de3b8c02f70b99392959f7f4aa56a1dbc9a08c5b