WebNN: Add the IDL definition of Softsign and five element-wise unary operators
This CL extracts a generic function of building the unary operator,
the output tensor of the unary operator must be the same type and
dimensions as its input tensor to use this help function.

This CL also implements the softsign and five element-wise unary IDL in
the MLGraphBuilder that be built and connected with input and output
operands. Or it can be built as an MLActivation.

Bug: 1273291
Change-Id: I7b81268a2c2adb89184ca32b786aa8d68a1b8e65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4987261
Reviewed-by: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Commit-Queue: Junwei Fu <junwei.fu@intel.com>
Cr-Commit-Position: refs/heads/main@{#1220695}
NOKEYCHECK=True
GitOrigin-RevId: 8c61b6e7fef549d00448f45fb2deb9d01f798b6f