Properly adjust StyleScope information when moving rules to diff rulesets.
When copying RuleData from a normal ruleset to a diff ruleset,
we would not adjust their position; this would give an entirely
wrong position, which would then crash as we tried finding the
appropriate StyleScope to test against.

Properly adjust the position through a helper (expanding on
the one we used for Bloom filter hashes), and add the StyleScope
to the proper intervals.

Fixed: 1499277

Change-Id: I6d99809ac137d644cec8dc2d4f3ccf863c32ef0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5001880
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220850}
NOKEYCHECK=True
GitOrigin-RevId: 70aa913c403de349711d2d3c3932bffa3d87b618