[A11y] Check aria-label to include elements with role="none" in AX tree
This is a reland of commit 1242b06ee200dd63b852956a67a6b3f7dae07ca4.
The original commit was reverted by LUCI due to test failures:
https://chromium-review.googlesource.com/c/chromium/src/+/5009239

The original description:
When an element is focusable or has a global ARIA property
and its role is set to 'none' or 'presentation', it should be
exposed in the accessibility tree with its native role.

However, the bug report indicates that Chrome does not take into
account the presence of the global ARIA property(aria-label).

This CL allows DoesUndoRolePresentation() to check for the existence
of the aria-label attribute.

Bug: 1466410
Test: Test: All/DumpAccessibilityTreeTest.AccessibilityLabelWithPresentationalChild/blink*
Change-Id: I525e9aa33ca04c1bcab4928ce770600ff05e80bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5008920
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1220903}
NOKEYCHECK=True
GitOrigin-RevId: c98d31d4474a0beb59225a6201c3993051e65286