Make the CSS math operation nested after distributing a multiplication over a sum
Due to the fact that CSS math trees are binary in Blink,
when we distribute a multiplication over a sum:
A - B * (C + D)
We don't know about the minus: A - ..., as
B * (C + D)
is the right-hand side of the operation A - ....
So, when we distribute B, we end up with
A - B * C + B * D,
which is incorrect.
Instead, leave the distributed operation nested, e.g.:
A - (B * C + B * D).

Note: The serialization code will not add unneeded parentheses:
(A + B) -> A + B.

Fixed: 1499774
Change-Id: If3158af7b2310b48ef6d5e2b0f3f8ecd37f0eb79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5009859
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1221554}
NOKEYCHECK=True
GitOrigin-RevId: fc0fe5d3ff4ea1a5d49ccdcdd5f3e3ca743b59ea