Speculative fix for crash when emulating forced colors via DevTools
This CL attempts to fix a crash that occurs when the DevTools feature
is used to emulate forced colors. The crash is caused by an
inappropriate change of the key from the renderer, which leads to a
call to a Windows OS method, GetSysColor(). The fix removes the key
change and uses the custom creation function that was already
implemented for the emulation. The crash could not be reproduced
locally and will be followed up once this lands in canary to confirm
that the fix works.

Bug: 1500381
Change-Id: I5ba9c29baad7d76bf79afe91eaecf8f9ebbed49d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5013943
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1221806}
NOKEYCHECK=True
GitOrigin-RevId: 0beb9ed81e28019956b8bd6f6e4f011ea9c39bfd