[FedCM] Repurpose ConfirmIdpLogin into ClickDialogButton
ConfirmIdpLogin is a webdriver command to click on the "Continue" button
on the FedCM mismatch dialog. To make it more extensible to other
buttons on the FedCM dialog, this patch
1. Renames ConfirmIdpLogin into ClickDialogButton
2. Passes DialogButton as a parameter to ClickDialogButton

In a follow-up patch, we will add support for buttons on the FedCM error
dialog.

Spec PR
https://github.com/fedidcg/FedCM/pull/510

WPT PR
https://github.com/web-platform-tests/wpt/pull/43002

Bug: 1499341
Change-Id: Iba12a1581dbe4236525a7d84b5f7ce11b95711f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5003255
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Zachary Tan <tanzachary@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1221852}
NOKEYCHECK=True
GitOrigin-RevId: e5e42e7e77c8d807ef3025e0841c874e2f9dc1a7