[iOS blink] Add kPreventResizingContentsForTesting flag...
for content_browsertests and web-platform-tests.

In a previous CL [1], the default size for tests was changed from
800x600 to 980x735 due to the issue with shrinking controls when
tests didn't provide a viewport meta.
To enable this change only for content_browsertests,
- Add switch |kPreventResizingContentsForTesting|
- Apply |kPreventResizingContentsForTesting| only for
  content_browsertests

For web-platform-tests, the recommended viewport size for web-platform-tests as 800x600 for conciseness, as stated in [3].
When rebaselining expectations for iOS in web-platform-tests,
it became concerns that a significant number of expectations generated
due to this viewport size difference.

Therefore, it would be the best for ios web platform tests to
have 800x600 viewport size. However, the problem is that most tests do not provide a viewport meta as well, which leads a scrollbar appearing with 800x600 default size due to ViewportStyle::kMobile, as seen in the code [2].

To address this for web-platform-tests,
- Sets ViewportStyle::kDefault for web-platform-tests

To reuse the platform specific expectations for ios,
- Sets "MacPort" as a fallback for "iOSPort" in web-platform-tests.

[1] [ios] Fix form controls tests (https://crrev.com/c/4615623)
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/css/resolver/viewport_style_resolver.cc;drc=719c18366c0226fa751f38693b00f822d1b7ea9b;l=83
[3] https://chromium.googlesource.com/chromium/src/+/main/docs/testing/web_tests_tips.md#concise

Bug: 1421239
Change-Id: Idd90db6152f5dbf1b261f386c9bea401f17c6bba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4991782
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Minju Kim <mkim@igalia.com>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1222023}
NOKEYCHECK=True
GitOrigin-RevId: 9d04730b1354626697d839f9c08d5ca2030b4b37