Fix BackgroundURLLoader's issue in handling redirect
Currently there is a bug in BackgroundURLLoader that if a request
was redirected, and the request is canceld by destructing the
BackgroundURLLoader before Context::OnReceivedRedirect() is called,
FollowRedirectCallback is destructed in the main thread. The
FollowRedirectCallback was created in the backgroud thread by
ResourceRequestSender::OnReceivedRedirect(), and it must be destructed
in the background thread. When DCHECK is enable,
DCHECK_CALLED_ON_VALID_SEQUENCE() in ThreadCheckingCallbackWrapper's
destructor fails.

To fix this issue, this CL changes RequestClient::OnReceivedRedirect()
to wrap the FollowRedirectCallback with a OnTaskRunnerDeleter as
a std::unique_ptr<FollowRedirectCallback, base::OnTaskRunnerDeleter>.
So the FollowRedirectCallback must be destructed in the background
thread.

Bug: 1379780
Change-Id: Iea7d4d1f17b137990d2c41709b497c7c034d5c47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5016226
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1222145}
NOKEYCHECK=True
GitOrigin-RevId: 1f5af861af63510fb30d8188cfa6a6ed8e87703a