PlzDedicatedWorker: Add the field trial testing config.
For ease of detecting changes that do not work well with
PlzDedicatedWorker, let me add the field trial testing config
with the PlzDedicatedWorker enabled.

We followed the past attempt to update the field trial testing config:
https://chromium.googlesource.com/chromium/src/+/c8900237101eff877c5eee2c1b17928c7f1abab9
i.e.
- move expectation files in /web_tests/virtual/plz-dedicated-worker
to the non-virtual directory.
- delete dedicated-worker-service-worker-interception.html that is not
spec-compatible behavior which was added at
https://chromium-review.googlesource.com/c/chromium/src/+/1059979
- delete dedicated-worker-reporting.html because COEP reports are sent
to worker's ReportingObserver API after PlzDedicatedWorker. Before
PlzDedicatedWorker, reports are sent to ancestor frame's
ReportingObserver but it's not spec compatible.
- delete use-counter-out-of-scope-worker.html, which was used for
measuring the impact of PlzDedicatedWorker ship for intent-to-ship.
- update "resource.type: other" to "resource.type: script".
- delete/merge "virtual/plz-dedicated-worker" expectation lines in
TestExpectations.

Removal of the "virtual/plz-dedicated-worker" directory will be
done in the future CL.

Bug: 906991
Change-Id: I89790cc16d5ed38880434811ae963886a3471087
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4983030
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1222215}
NOKEYCHECK=True
GitOrigin-RevId: 38605a0d7b8bfe6419da09257a4f92195908dee6