Notify Renderer of Surface Eviction and Determine Locked Memory
The ActiveTree of a Renderer can lock viz::TransferableResources against
destruction.

When we Evict a viz::Surface we want to reclaim all memory. Currently
these locks lead to some resources being held until the Renderer is
destroyed.

This change adds a new signal to CompositorFrameSinkClient to notify
them that a viz::Surface was evicted. We are currently connecting this
to viz::ClientResourceProvider so we can detect the locked resources.

In a follow-up we will begin releasing these, dependent on the
allocation source.

Bug: 1449273
Change-Id: I86411f68eeb33d4e4fc6aaa815a7a4d596b21d49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4956873
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1222498}
NOKEYCHECK=True
GitOrigin-RevId: f5251985f74a48c37a6df31be38de4de2e9bd202