Account for document.head() returning null
It is possible for document.head() to return null. Most other call sites
check for this. Add the missing check here. This was causing UBSan
failures in DocumentLoaderTest.

There seem to be a few of these in Blink. Blink likes to use references
when it believes fields are non-null. Tracking non-null pointers is
good, but Blink seems to constantly dereference its nullable pointers
without null checks. The point of categorizing things is to force you to
check for the null on the transition from nullable to non-nullable, but
Blink's coding patterns seem not to do this.

Bug: 1394755
Change-Id: Ie72d47a13ca56b0d91d303c77ecdd2a3cfbff18c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017782
Auto-Submit: David Benjamin <davidben@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1222731}
NOKEYCHECK=True
GitOrigin-RevId: 34bcf67e4fbd83f2f0294d9a94acd406eb34f6f6