Rename SpeculationInjectionWorld to SpeculationInjectionType
As part of the AutoSpeculationRules feature, we want to track browser-injected auto speculation rules. The existing machinery used for tracking whether a speculation rule is from an isolated world script, or not, is useful for this. But the name is not general enough.

Bug: 1472970
Change-Id: Ib80b7411910211ab02b242d2e40806c932ba481e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5013491
Auto-Submit: Domenic Denicola <domenic@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Domenic Denicola <domenic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1222752}
NOKEYCHECK=True
GitOrigin-RevId: ff44ecf98ab5b73483a55ade769ba3bab49455bc